% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA data}
\usage{
eq_clean_data(raw)
}
\arguments{
\item{raw}{Raw dataframe obtained after reading NOAA Significant Earthquake Database tab-separated file}
}
\value{
The clean dataframe after creating DATE column and transforming LATITUDE/LONGITUDE columns
}
\description{
Clean NOAA Significant Earthquake Database and store it into a dataframe
}
\details{
The clean data frame should have the following:
\enumerate{
   \item DATE column created by uniting the year, month, day and converting it to the Date class
   \item LATITUDE and LONGITUDE columns converted to numeric class
}
}
\examples{
raw <- readr::read_tsv('signif.txt')
clean <- eq_clean_data(raw)

}
