% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Earthquake labels}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{NOAA earthquake data in a clean dataframe (after a call to \code{eq_clean_data})}
}
\value{
A character vector with the labels
}
\description{
Create labels for earthquake data to be included in leaflet maps
}
\details{
It takes the dataset as an argument and creates an HTML label that
can be used as the annotation text in the leaflet map. This function
puts together a character string for each earthquake that will show
the cleaned location (as cleaned by the eq_location_clean()
function), the magnitude (EQ_PRIMARY), and the total number of
deaths (TOTAL_DEATHS), with boldface labels for each
("Location", "Total deaths", and "Magnitude"). If an earthquake is
missing values for any of these, both the label and the value should
be skipped for that element of the tag
}
\examples{
data \%>\%
   filter(COUNTRY == "MEXICO" & year(DATE) >= 2000) \%>\%
   mutate(popup_text = eq_create_label(.)) \%>\%
   eq_map(annot_col = "popup_text")

}
