% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean location}
\usage{
eq_location_clean(raw)
}
\arguments{
\item{raw}{Raw dataframe obtained after reading NOAA Significant Earthquake Database tab-separated file}
}
\value{
The clean dataframe after transforming LOCATION_NAME column
}
\description{
Clean location from NOAA Significant Earthquake Database and return it as a vector
}
\details{
Cleans the LOCATION_NAME column by stripping out the country name (including the colon) and converts names
to title case (as opposed to all caps). This function should be applied to the raw data to produce a cleaned
up version of the LOCATION_NAME column.
}
\examples{
raw <- readr::read_tsv('signif.txt')
location_clean <- eq_location_clean(raw)

}
