% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_map}
\alias{eq_map}
\title{Earthquake maps}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{NOAA earthquake data in a clean dataframe (after a call to \code{eq_clean_data})}

\item{annot_col}{The column that will be taken for anotaion of the earthquakes}
}
\value{
The leaflet map with the earthquake data
}
\description{
Generate maps based on NOAA earthquake data
}
\details{
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates each point with
a pop up window containing annotation data stored in a column of the data frame.
The user should be able to choose which column is used for the annotation in the pop-up
with a function argument named annot_col. Each earthquake should be shown with a circle,
and the radius of the circle should be proportional to the earthquake's
magnitude (EQ_PRIMARY).
}
\examples{
library(dplyr)
library(lubridate)
earthquake_data \%>\%
  filter(COUNTRY == "MEXICO" & year(DATE) >= 2000) \%>\%
  eq_map(annot_col = "DATE")

}
