% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{top_earthquakes}
\alias{top_earthquakes}
\title{Filter earthquakes by magnitude}
\usage{
top_earthquakes(data, n_max, date_min = -Inf, date_max = Inf)
}
\arguments{
\item{data}{Earthquake data}

\item{n_max}{Max number of earthquakes to pick}

\item{date_min}{Minimum date}

\item{date_max}{Maximum date}
}
\value{
A dataframe with the filtered data
}
\description{
Pick the top n_max earthquakes acoording to their magnitude.
}
\examples{
top_10 <- top_earthquakes(earthquake_data_raw, 10)

}
