% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{build_label}
\alias{build_label}
\title{Build map label}
\usage{
build_label(tag, content)
}
\arguments{
\item{tag}{preffix, to be in bold. E.g. Location:}

\item{content}{values to be added after the tag}
}
\value{
The labels to be displayed in the popup
}
\description{
It builds the individual labels for the popup in the map.
}
\details{
In the case of one value to be NA, the corresponding generated label is ''
}
